replace old sort component
