+++ ui/marketplace/utils.ts
@@ -2,6 +2,14 @@ import type { NextRouter } from 'next/router';
 
 import getQueryParamString from 'lib/router/getQueryParamString';
 import removeQueryParam from 'lib/router/removeQueryParam';
+import type { TOption } from 'ui/shared/sort/Option';
+
+export type SortValue = 'security_score';
+
+export const SORT_OPTIONS: Array<TOption<SortValue>> = [
+  { title: 'Default', id: undefined },
+  { title: 'Security score', id: 'security_score' },
+];
 
 export function getAppUrl(url: string | undefined, router: NextRouter) {
   if (!url) {
