+++ ui/nameDomains/utils.ts
@@ -1,12 +1,12 @@
 import type { EnsLookupSorting } from 'types/api/ens';
 
 import getNextSortValueShared from 'ui/shared/sort/getNextSortValue';
-import type { Option } from 'ui/shared/sort/Sort';
+import type { TOption } from 'ui/shared/sort/Option';
 
 export type SortField = EnsLookupSorting['sort'];
 export type Sort = `${ EnsLookupSorting['sort'] }-${ EnsLookupSorting['order'] }`;
 
-export const SORT_OPTIONS: Array<Option<Sort>> = [
+export const SORT_OPTIONS: Array<TOption<Sort>> = [
   { title: 'Default', id: undefined },
   { title: 'Registered on descending', id: 'registration_date-DESC' },
   { title: 'Registered on ascending', id: 'registration_date-ASC' },
